Beginning to allow user save their old patterns
Began working on letting a user save their pattern by using the Upload
Array and Download Array capabilities in the GalilWidgets interface. For
the most part it is good except after we have saved the arrays to the
computer the rotation[] array element always has zero but it must have
the size of the posC[] array.